feat: add version package to handle module version and commit
